requirements.txt
google-adk==1.13.0
litellm==1.77.0
reportlab==4.4.3